use GBLCTL instead of RGBLCTL/XGBLCTL
RGBLCTL and XGBLCTL always read back the last value written to GBLCTL
(directly, or implicitly by writing to RGBLCTL or XGBLCTL).  To poll for
acknowledgement, GBLCTL itself *must* be used.

Writing to RGBLCTL/XGBLCTL restricts the write to affect only the bits
affecting the receive section or transmit section respectively. This
is used to allow the receive and transmit sections to be managed by
different cores in parallel without mutex or race condition.  For this
firmware this feature is irrelevant, hence it's simpler to just always
use GBLCTL itself.