change 'static' functions in xenomai_wraps.h to 'inline'
I am reusing xenomai_wraps.h to implement a pipe and queue without starting
a separate task for each of them.  When including xenomai_wraps.h, I
get the following warning:

/home/mlang/Modular/Bela/include/xenomai_wraps.h: At global scope:
/home/mlang/Modular/Bela/include/xenomai_wraps.h:137:12: warning: ‘int create_and_start_thread(pthread_t*, const char*, int, int, void* (*)(void*), void*)’ defined but not used [-Wunused-function]
 static int create_and_start_thread(pthread_t* task, const char* taskName, int priority, int stackSize, pthread_callback_t* callback, void* arg)
            ^~~~~~~~~~~~~~~~~~~~~~~

These really shouldn't be static, inline is the correct way to go for
code in a C++ header.

(while at it, fix a double semicolon.)