IDE: ensure the horizontal size of the editor is the one that is actually visible.
Closes https://github.com/BelaPlatform/Bela-private/issues/256.
Also, as the vertical scrollbar is expected to be always hidden
(https://github.com/BelaPlatform/Bela-private/issues/217), at least we
make this clear in the comments. Arguably a better fix would be to show
it only on hover or scroll.
At the same time, do better for the transition so that the horizontal
scrollbar resizes smoothly (fixup for 7ee305a5bd2399df04ac6afd7818be3977636a1d).