set BBB_NETWORK_TARGET_FOLDER after check_board_alive_and_set_date
I noticed that rsync failed and realized that it was because my board
was showing up at 192.168.6.2 instead of the default BBB_ADDRESS.

check_board_alive_and_set_date updates BBB_ADDRESS to the correct
address, but because BBB_NETWORK_TARGET_FOLDER was being set before
running this check, it would use the default BBB_ADDRESS.

We need to set the network target folder *after* running this check,
or the default BBB_ADDRESS gets copied into the target folder,
and the change when checking that the board is alive does
not propagate. This causes rsync in uploadBuildRun to fail.