RTAudio: start codec _after_ PRU has started
This further reduces the startup pop the first time the program runs
after boot, at least with the I2c_Codec. It does so by starting the DAC
after the PRU and McASP clocks have been started. It is unclear to me
why it behaves differently after the first start: there may be
something that gets initialised in the PRU code and doesn't get
cleared...

This was tested only on BelaMini.